fix imagining

+++ app/src/main/java/com/pimenov/ozon/presentation/mapper/ProductMapper.kt
@@ -6,7 +6,7 @@ import com.pimenov.ozon.presentation.viewObject.ProductInListPresentation
 import com.pimenov.ozon.presentation.viewObject.ProductPresentation
 
 fun ProductInList.toPresentation() : ProductInListPresentation =
-    ProductInListPresentation(guid, image, name, price, rating, isFavorite, isInCart)
+    ProductInListPresentation(guid, image, name, price, rating.toFloat(), isFavorite, isInCart)
 
 fun Product.toPresentation() : ProductPresentation =
     ProductPresentation(guid, name, price, description, rating, isFavorite, isInCart, images, weight, count, availableCount, additionalParams)
